Actions insert elements at cursor position
Insertion of elements via the buttons in the Actions section now happens at the cursor position. If the textarea does not have a cursor, the insertion happens at the end of the content. The operation can be undone using CTRL+Z, although to enable this feature it was necessary to adopt a non-future-proof solution, namely using the deprecated "execCommand" method.